Refactor color scale
